refactor -> move AlertMessage to alert package
