+++ ChatterClient/src/main/java/com/codecool/chatter/view/box/CreateRoomForm.java
@@ -1,6 +1,6 @@
 package com.codecool.chatter.view.box;
 
-import com.codecool.chatter.view.AlertMessage;
+import com.codecool.chatter.view.alert.AlertMessage;
 import com.codecool.chatter.view.interactive.ButtonView;
 import com.codecool.chatter.view.interactive.InputField;
 import javafx.event.EventHandler;
